working taxonomy dictionary!
This version of lcplants.py is a bit of a mess, but I DID finally figure out how to build the taxonomy tree. Right now the code for making the dictionary is doubled up with the code producing kumu elements-- I'll untangle it later, but wanted to save this version before I start re-assembling it.

The kumu map it produces is viewable here, although I'm messing with the clustering code so it's also messy. I'm uploading my kumu settings too.

https://kumu.io/violetmadder/taxonomy03#taxonomy

(this link will eventually break when I have better files and delete the old ones)